requirements.txt
tweepy
google-generativeai
requests
certifi
feedparser
re
random
subprocess
os